+++ lib/src/builder/widget_repository_builder.dart
@@ -1,16 +1,16 @@
 import 'package:build/build.dart';
 import 'package:glob/glob.dart';
-import 'package:path/path.dart' as path;
 import 'package:logging/logging.dart';
+import 'package:path/path.dart' as path;
 import 'package:yet_another_layout_builder/src/builder/styles_collector.dart';
 
+import 'class_finders.dart';
 import 'found_items.dart';
+import 'generator/code_generator.dart';
 import 'path_matcher.dart';
 import 'progress_collector.dart';
-import 'class_finders.dart';
 import 'widget_helpers.dart';
 import 'xml_analyzer.dart';
-import 'generator/code_generator.dart';
 
 Builder widgetRepoBuilder(BuilderOptions options) =>
     WidgetRepositoryBuilder(options);
@@ -76,7 +76,8 @@ class WidgetRepositoryBuilder implements Builder {
     List<Resolvable> allConsts = [];
     List<Resolvable> allWidgets = [];
     for (var widget in widgets) {
-      allWidgets.add(Resolvable(widget.name, widget.attributes, null, ""));
+      allWidgets.add(Resolvable(
+          widget.name, widget.attributes, widget.designatedCtrName, ""));
       collectConst(widget.constItems, allConsts);
     }